move pointcut parsing implementation to dedicated directory
