+++ packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/Pointcut/PointcutInterfaceExpression.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Ecotone\Messaging\Handler\Processor\MethodInvoker;
+namespace Ecotone\Messaging\Handler\Processor\MethodInvoker\Pointcut;
 
 use Ecotone\Messaging\Handler\ClassDefinition;
 use Ecotone\Messaging\Handler\InterfaceToCall;
+use Ecotone\Messaging\Handler\Processor\MethodInvoker\PointcutExpression;
 
 class PointcutInterfaceExpression implements PointcutExpression
 {
