+++ packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/Pointcut/PointcutOrExpression.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Ecotone\Messaging\Handler\Processor\MethodInvoker;
+namespace Ecotone\Messaging\Handler\Processor\MethodInvoker\Pointcut;
 
 use Ecotone\Messaging\Handler\InterfaceToCall;
+use Ecotone\Messaging\Handler\Processor\MethodInvoker\PointcutExpression;
 
-class PointcutAndExpression implements PointcutExpression
+class PointcutOrExpression implements PointcutExpression
 {
     public function __construct(private PointcutExpression $left, private PointcutExpression $right)
     {
@@ -13,6 +14,6 @@ class PointcutAndExpression implements PointcutExpression
     public function doesItCutWith(array $endpointAnnotations, InterfaceToCall $interfaceToCall): bool
     {
         return $this->right->doesItCutWith($endpointAnnotations, $interfaceToCall)
-            && $this->left->doesItCutWith($endpointAnnotations, $interfaceToCall);
+            || $this->left->doesItCutWith($endpointAnnotations, $interfaceToCall);
     }
 }
