+++ packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/Pointcut/PointcutRegexExpression.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Ecotone\Messaging\Handler\Processor\MethodInvoker;
+namespace Ecotone\Messaging\Handler\Processor\MethodInvoker\Pointcut;
 
 use Ecotone\Messaging\Handler\InterfaceToCall;
+use Ecotone\Messaging\Handler\Processor\MethodInvoker\PointcutExpression;
 
 class PointcutRegexExpression implements PointcutExpression
 {
